requirements.txt
appdirs==1.4.3
APScheduler==3.8.1
certifi==2018.11.29
chardet==3.0.4
Click==7.0
cycler==0.10.0
distlib==0.3.0
dnspython==1.15.0
dnspython3==1.15.0
filelock==3.0.12
idna==2.8
importlib-metadata==1.6.0
joblib==0.17.0
Mako==1.1.2
MarkupSafe==1.1.1
pyaml==21.10.1
pycparser==2.20
pyparsing==2.4.7
python-dateutil==2.8.1
python-editor==1.0.4
pytz==2020.4
pytz-deprecation-shim==0.1.0.post0
PyYAML==6.0
requests==2.21.0
six==1.14.0
soupsieve==2.2.1
threadpoolctl==2.1.0
tzdata==2021.5
tzlocal<3.0
urllib3==1.24.1
virtualenv==20.0.15
zipp==3.1.0
